multi class file division
